pyproject.toml
[project]
name = "examen-p2"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard] (>=0.136.1,<0.137.0)",
    "sqlmodel (>=0.0.38,<0.0.39)",
    "pwdlib[argon2] (>=0.3.0,<0.4.0)",
    "pyjwt (>=2.12.1,<3.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)"
]

[tool.poetry]
packages = [{include = "examen_p2", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
